requirements.txt
PyMuPDF
easyocr
frontend
pypdf2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
flask == 2.3.3
flask_cors
pyjwt